+++ hello_world/src/main.rs
@@ -1,5 +1,6 @@
 use std::env;
-use std::collections::HashMap;
+pub mod fizzbuzz;
+pub use self::fizzbuzz::*;
 
 
 struct Color {
@@ -250,14 +251,14 @@ mod tests {
         }
     }
 
-    #[test]
+    // #[test]
 
-    fn test_fizz_buzz() {
-        assert_eq!(determine_fizz_buzz(4), format!("{}", 4));
-        assert_eq!(determine_fizz_buzz(3), "fizz");
-        assert_eq!(determine_fizz_buzz(5), "buzz");
-        assert_eq!(determine_fizz_buzz(15), "fizz buzz");
-    }
+    // fn test_fizz_buzz() {
+    //     assert_eq!(determine_fizz_buzz(4), format!("{}", 4));
+    //     assert_eq!(determine_fizz_buzz(3), "fizz");
+    //     assert_eq!(determine_fizz_buzz(5), "buzz");
+    //     assert_eq!(determine_fizz_buzz(15), "fizz buzz");
+    // }
 
     // #[test]
 
@@ -328,14 +329,4 @@ fn create_array(elements: i32,len: usize) -> Vec<i32> {
 fn increment_value(x: &mut i32) -> &mut i32 {
     *x += 1;
     x
-}
-
-fn determine_fizz_buzz(input: i64) -> String {
-    let result = match (input%3, input%5) {
-        (0,0) => String::from("fizz buzz"),
-        (0,_) => String::from("fizz"),
-        (_,0) => String::from("buzz"),
-        (_,_) => format!("{}", input),
-    };
-    result
 }
